Reorganize modules and rename Midi to MidiInputPlugin
