+++ examples/minimal.rs
@@ -1,9 +1,9 @@
 use bevy::prelude::*;
-use bevy_midi::Midi;
+use bevy_midi::MidiInputPlugin;
 
 fn main() {
     App::new()
         .add_plugins(MinimalPlugins)
-        .add_plugin(Midi)
+        .add_plugin(MidiInputPlugin)
         .run();
 }
